v32sim changes:
  * offset display issue: uint8_t changed to uint32_t for value in
    put_word() in word.c

The observed  discrepancy in the  displayed IP  offset turns out  to have
been caused by an improperly  sized intermediate variable: in put_word(),
to accommodate  immediate data,  I stored  the IP  register value  into a
variable, so if the immediate flag were set I could increment it by one.

The issue was, I accidentally made  that variable a uint8_t, which caused
all manner  of data to  be lost. Changing it  to a uint32_t  resolved the
issue.

+++ word.c
@@ -64,7 +64,7 @@ void      put_word (uint32_t  word, uint8_t  flag)
     // Declare and initialize local variables
     //
     int32_t   index     = 0;
-    uint8_t   value     = REG(IP);
+    uint32_t  value     = REG(IP);
     uint32_t  mask      = 0xFF000000;
 
     if (FLAG_IMMEDIATE == (flag & FLAG_IMMEDIATE))
